Added regression testing using newman.

+++ server/controllers/playersController.js
@@ -1,3 +1,5 @@
+import Sequelize from "sequelize";
+
 import { Player, ManagementRelationship } from "../models/index";
 import playerSerializer from "../serializers/playerSerializer";
 
@@ -54,20 +56,35 @@ const create_player = async (req, res) => {
 
 const delete_player = async (req, res) => {
   try {
-    const player_records_destroyed = await Player.destroy({
+    const playerUUID = req.params.id;
+    const adminUUID = req.body.user.id;
+    const Op = Sequelize.Op;
+
+    const managementRecord = await ManagementRelationship.findOne({
       where: {
-        id: req.params.id
+        adminUUID,
+        [Op.and]: {
+          playerUUID
+        }
       }
     });
 
-    if (player_records_destroyed == 1) {
-      const relationship_records_destroyed = await ManagementRelationship.destroy(
-        {
-          where: {
-            playerUUID: req.params.id
-          }
+    if (managementRecord) {
+      const player_records_destroyed = await Player.destroy({
+        where: {
+          id: req.params.id
         }
-      );
+      });
+
+      if (player_records_destroyed == 1) {
+        const relationship_records_destroyed = await ManagementRelationship.destroy(
+          {
+            where: {
+              playerUUID: req.params.id
+            }
+          }
+        );
+      }
     } else {
       throw "Error in record destruction";
     }
